+++ app/(tabs)/products.tsx
