+++ app/(tabs)/stock-manager.tsx
