+++ app/(tabs)/stock-summary.tsx
