+++ app/(tabs)/warehouse.tsx
