+++ app/_layout.tsx
@@ -1,6 +1,6 @@
 import { DarkTheme, DefaultTheme, ThemeProvider } from '@react-navigation/native'
 import { useFonts } from 'expo-font'
-import { Redirect, Tabs } from 'expo-router'
+import { Redirect, Stack, Tabs } from 'expo-router'
 import * as SplashScreen from 'expo-splash-screen'
 import React, { useEffect } from 'react'
 import 'react-native-reanimated'
@@ -16,44 +16,6 @@ import { Easing } from 'react-native'
 
 // Prevent the splash screen from auto-hiding before asset loading is complete.
 SplashScreen.preventAutoHideAsync()
-
-const tabs = [
-  {
-    name: 'profile',
-    title: 'Perfil',
-    showRootHeader: true,
-    iconName: 'person',
-  },
-  {
-    name: 'stock-manager',
-    title: 'Administrar Stock',
-    showRootHeader: true,
-    iconName: 'archive',
-  },
-  {
-    name: 'index',
-    title: 'Inicio',
-    iconName: 'home',
-  },
-  {
-    name: 'warehouse',
-    title: 'Depósitos',
-    iconName: 'warehouse',
-  },
-  {
-    name: 'products',
-    title: 'Productos',
-    iconName: 'cube',
-  },
-]
-
-const invisibleTabs = [
-  { name: '+not-found' },
-  { name: 'login' },
-  { name: 'register' },
-  { name: 'stock-summary', title: 'Resumen Stock', showRootHeader: true },
-  { name: 'reports', title: 'Reportes', showRootHeader: true },
-]
 
 export default function RootLayout() {
   const colorScheme = useColorScheme()
@@ -79,86 +41,11 @@ export default function RootLayout() {
       <AuthProvider>
         <QueryClientProvider client={queryClient}>
           <GestureHandlerRootView>
-            <Tabs
-              screenOptions={({ route }) => ({
-                tabBarHideOnKeyboard: true,
-                headerShown: false,
-                tabBarVisibilityAnimationConfig: {
-                  hide: {
-                    config: { duration: 150 },
-                    animation: 'timing',
-                  },
-                },
-                tabBarStyle: {
-                  display: ['login', 'register'].includes(route.name) ? 'none' : 'flex',
-                  position: 'absolute',
-                  paddingBottom: 10,
-                  paddingTop: 10,
-                  height: 65,
-                  width: '90%',
-                  left: '5%',
-                  borderTopRightRadius: 16,
-                  borderTopLeftRadius: 16,
-                  backgroundColor:
-                    colorScheme === 'dark'
-                      ? DarkTheme.colors.background
-                      : DefaultTheme.colors.background,
-                },
-              })}
-            >
-              {tabs.map(({ name, title, iconName, showRootHeader = false }, index) => (
-                <Tabs.Screen
-                  key={index}
-                  name={name}
-                  options={{
-                    unmountOnBlur: true,
-                    title: title,
-                    header: (props: any) => <PageHeader {...props} back={undefined} />,
-                    headerShown: showRootHeader,
-                    tabBarIcon: ({ focused }) =>
-                      iconName !== 'warehouse' ? (
-                        <Ionicons
-                          name={iconName as keyof typeof Ionicons.glyphMap}
-                          size={24}
-                          color={
-                            focused
-                              ? Colors.primary[600]
-                              : colorScheme === 'dark'
-                                ? Colors.gray[300]
-                                : Colors.gray[600]
-                          }
-                        />
-                      ) : (
-                        <FontAwesome6
-                          name={iconName}
-                          size={20}
-                          color={
-                            focused
-                              ? Colors.primary[600]
-                              : colorScheme === 'dark'
-                                ? Colors.gray[300]
-                                : Colors.gray[600]
-                          }
-                        />
-                      ),
-                  }}
-                />
-              ))}
-              {invisibleTabs.map(({ name, title = '', showRootHeader }, index) => (
-                <Tabs.Screen
-                  key={index}
-                  name={name}
-                  options={{
-                    title: title,
-                    href: null,
-                    header: showRootHeader
-                      ? (props: any) => <PageHeader {...props} back={undefined} />
-                      : () => <></>,
-                    headerShown: showRootHeader,
-                  }}
-                />
-              ))}
-            </Tabs>
+            <Stack>
+              <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
+              <Stack.Screen name="login" options={{ headerShown: false }} />
+              <Stack.Screen name="register" options={{ headerShown: false }} />
+            </Stack>
           </GestureHandlerRootView>
         </QueryClientProvider>
       </AuthProvider>
