+++ app/index.tsx
@@ -1,78 +1,8 @@
-import CardLineChart from '@/components/charts/CardLineChart'
-import Container from '@/components/Container'
-import IconCard from '@/components/IconCard'
-import Typography from '@/components/Typography'
-import { useAuthorizedUser } from '@/hooks/useUser'
-import { router } from 'expo-router'
 import React from 'react'
-import { View } from 'react-native'
-import { ScrollView } from 'react-native-gesture-handler'
+import { Redirect } from 'expo-router'
 
-const Dashboard = () => {
-  const { user } = useAuthorizedUser()
-  const lineData = [
-    { value: 0 },
-    { value: 20 },
-    { value: 8 },
-    { value: 40 },
-    { value: 12 },
-    { value: 30 },
-    { value: 90 },
-    { value: 75 },
-  ]
-  return (
-    user && (
-      <Container style={{ height: '100%' }}>
-        <ScrollView showsVerticalScrollIndicator={false} contentContainerStyle={{}}>
-          <View
-            style={{
-              marginTop: 70,
-              padding: 16,
-              justifyContent: 'center',
-              alignItems: 'center',
-              gap: 20,
-              paddingBottom: 100,
-            }}
-          >
-            <Typography variant="h4">Bienvenido {user.user}</Typography>
-            <View
-              style={{
-                flexDirection: 'row',
-                justifyContent: 'space-between',
-                gap: 20,
-              }}
-            >
-              <IconCard
-                icon="cube-outline"
-                color={'primary'}
-                text="Productos"
-                onPress={() => {
-                  router.push('/products')
-                }}
-              />
-              <IconCard
-                icon="warehouse"
-                color={'gray'}
-                text="Depósitos"
-                onPress={() => {
-                  router.push('/warehouse')
-                }}
-              />
-            </View>
-            <CardLineChart data={lineData} />
-            <IconCard
-              icon="bar-chart-outline"
-              color={'primary'}
-              text="Reportes"
-              onPress={() => {
-                router.push('/reports')
-              }}
-            />
-          </View>
-        </ScrollView>
-      </Container>
-    )
-  )
+const index = () => {
+  return <Redirect href="/dashboard" />
 }
 
-export default Dashboard
+export default index
